+++ tests/test_runner.py
@@ -121,7 +121,7 @@ class TestVerdictToHookResponse:
 
     def test_warn_returns_warning_system_message(self) -> None:
         resp = runner.verdict_to_hook_response(self._rule("warn"), "mild issue", "warn")
-        assert resp["decision"] == "block"
+        assert resp["decision"] == "warn"
         assert "Warning" in resp["systemMessage"]
         assert "mild issue" in resp["systemMessage"]
 
@@ -220,7 +220,7 @@ class TestRunPipeline:
         resp = json.loads(out.strip())
         assert resp["decision"] == "block"
 
-    def test_null_classify_result_continues_to_next_rule(self, capsys) -> None:
+    def test_null_classify_result_continues_to_next_rule(self) -> None:
         hook_input = {
             "session_id": "null-result-session",
             "last_assistant_message": "A" * 200,
@@ -237,7 +237,7 @@ class TestRunPipeline:
                 runner._run()
         assert exc_info.value.code == 0
 
-    def test_text_below_min_length_skips_rule(self, capsys) -> None:
+    def test_text_below_min_length_skips_rule(self) -> None:
         hook_input = {
             "session_id": "short-text-session",
             "last_assistant_message": "short",
@@ -254,7 +254,7 @@ class TestRunPipeline:
         assert exc_info.value.code == 0
         mock_client.classify.assert_not_called()
 
-    def test_missing_rule_skips_silently(self, capsys) -> None:
+    def test_missing_rule_skips_silently(self) -> None:
         hook_input = {
             "session_id": "missing-rule-session",
             "last_assistant_message": "A" * 200,
